Update Square with onClick functionality
The Square class was not performing any actions when clicked on,
which is necessary for the game of tic-tac-toe. Therefore an onClick
function was added to the render of Square inside of the button tag
so that each Square would actually be treated like a button and perform
an action when clicked on. A constructor was also added to Square to
set the state of each Square for proper functionality of the buttons.

- this.props.value -> this.state.value