Refactor value of Square in renderSquare()
The value of Square inside of renderSquare() was set to just i,
which was used for incrementing the numbers that were previously
being displayed inside of the squares, but we do not need that
anymore. Instead this value was set to the squares variable inside
of the Board's state incrementally so that a null will be given to
each Square when rendered.