Add Board click handler
The Square class calls upons the onClick method from Board props,
but there is no method defined to actually handle the click, which
will cause an error when a square is clicked. So a handleClick() was
defined in the Board class to control the components when they are
ultimately clicked.