Refactor Square class with only render
THe Square class was previously defined as an extension of
React.component, but since we are only using render() for Square,
this is overcomplicating the function. Square was reduced to a
stateless functional component which only uses a render(). To
complete this all of the this.props are now just changed to props.