Update handleclick() to toggle player
The state is setup for changing the state/status of whether or not
X is the next player or not. However, the handleClick() is not setup
for toggling the boolean value of xIsNext aka the player. The
constant for squares inside of handleClick() was changed to be the
state of xIsNext depending on who the current player is, which would
then determine how this.setState acts, and thus who the next player
is going to be.