Add if/else conditional for status value
The value for the status constant would previously only display who
the next player is, but if someone has won the game there is no next
player and you want the game to display who won instead. Thus status
was changed to a let status ... conditional if/else where if there
is a determined winner that player will be displayed, else the
next player will be displayed.