Set initial state for Game
To store the history of the game, the state must be moved upwards
from Game to Board, just like the state was moved upwards from Board
to Square earlier. To begin the process the same constructor from
Board was added to Game with a variable of history saved in the
Game's state and within history is squares.