Refactor renderSquare() in Board
The renderSquare() in Board only contained a return statement, but
needed to take squares via props and have its own onClick()
specified by Game, therefore squares was defined as a constant of
this.props.squares. The value inside of Board's return was then
changed to equal just squares[i] and the onClick() was converted to
equal this.props.onClick(), removing the handleClick().