Add history and current to handleClick()
The same constants for history and current were added to handleClick()
so that the Game can properly handle all clicks of the game while
updating the squares and history. The function also now contains a
conditional for running the return depending on the outcome of
calculateWinner(). The this.setState() of the handleClick() function
was also changed to include the concatenation of the values of the
squares of the game and storing them in history.