Add move counter to Game render()
For the full effect of the game, we want to be able to view the
previous moves made within the game. To do this a constant of moves
was added that takes the history and iterates over it mapping each
'step' and 'move' . The move number is displayed unless there are
no previous moves, in whichc case the game displays 'Game Start'.
The moves are displayed on the page just below the status of the game.