Add unique move key to li elements
The methods implemented for moves raised a warning for having a unique key
for each li child of the moves list for identification. Each of the
moves already has a unique identifier which is their move number
titled 'move' so a key equaling move was added to the li element
that is in the return statement for moves.