Add stepNumber to inital state
The <a> inside of moves return calls upon a jumpTo method that has
not been defined yet. This method will need to take a stepNumber
and continuously update it with another step, but before completely
defining the jumpTo() a stepNumber must exist as part of the initial
state of the Game. Therefore a stepNumber of 0 was added into the
initial state of Game defined within its constructor.